Added support for building 64bit Windows images without floppy.
This patch makes the assumption that the ISO containing virtio
drivers is already cached.  The other assumption is that the
architecture is x86_64.  The install ISO is combined together
with an autounattend.xml in a new ISO.  The new ISO is uploaded
to cinder and is then attached to an instance as the cdrom.  The
ISO containing the drivers is also attached to the instance.